fix: share desktop command contracts for Docker web builds
