+++ desktop/src/shared/desktop-bridge-contracts.ts
@@ -1,5 +1,19 @@
 import type { DesktopAppInfo, DesktopStatusSnapshot } from '../main/types.js';
-import { blockedRemoteConnectionDestinationReason } from '@veritas-kanban/shared';
+import {
+  blockedRemoteConnectionDestinationReason,
+  DESKTOP_COMMAND_NAMES,
+  type DesktopCommandName,
+  type DesktopCommandSource,
+  type DesktopCommandDispatchRequest,
+  type DesktopCommandDispatchResult,
+} from '@veritas-kanban/shared';
+export {
+  DESKTOP_COMMAND_NAMES,
+  type DesktopCommandName,
+  type DesktopCommandSource,
+  type DesktopCommandDispatchRequest,
+  type DesktopCommandDispatchResult,
+} from '@veritas-kanban/shared';
 
 export const DESKTOP_REDACTED_VALUE = '[redacted]';
 export const DESKTOP_RESTART_CONFIRMATION = 'restart-local-server';
@@ -82,47 +96,6 @@ export interface DesktopSetupDiagnostics {
   generatedAt: string;
   checks: DesktopSetupDiagnosticCheck[];
   supportSnapshot: DesktopSupportSnapshot;
-}
-
-export const DESKTOP_COMMAND_NAMES = [
-  'new-task',
-  'open-onboarding',
-  'open-search',
-  'open-settings',
-  'open-command-center',
-  'reset-layout',
-  'import-data',
-  'export-data',
-  'create-backup',
-  'open-logs',
-  'restart-local-server',
-  'communication-health',
-  'show-diagnostics',
-  'create-debug-bundle',
-  'check-for-updates',
-  'download-update',
-  'install-update',
-  'test-notification',
-  'test-squad-webhook',
-  'copy-redacted-diagnostics',
-  'export-work-product',
-  'quit',
-] as const;
-
-export type DesktopCommandName = (typeof DESKTOP_COMMAND_NAMES)[number];
-export type DesktopCommandSource = 'renderer' | 'menu' | 'shortcut' | 'deep-link';
-
-export interface DesktopCommandDispatchRequest {
-  command: DesktopCommandName;
-  source?: DesktopCommandSource;
-  payload?: Record<string, unknown>;
-}
-
-export interface DesktopCommandDispatchResult {
-  command: DesktopCommandName;
-  accepted: boolean;
-  handledBy: 'desktop' | 'renderer' | 'unsupported';
-  message?: string;
 }
 
 export const DESKTOP_FILE_PICKER_PURPOSES = [
