+++ shared/src/types/index.ts
@@ -81,3 +81,5 @@ export * from './tool-control-plane.types.js';
 export * from './acp.types.js';
 export * from './workspace-execution-trust.types.js';
 export * from './phase-capability.types.js';
+
+export * from './desktop-commands.js';
